llm-compliance/requirements.txt
agentevals==0.0.9
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bert-score==0.3.13
blis==1.3.3
build==1.3.0
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.2.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.13
databricks-sdk==0.76.0
dataclasses-json==0.6.7
datasets==4.4.2
decorator==5.2.1
deepeval==3.7.6
detoxify==0.5.2
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
durationpy==0.10
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
execnet==2.1.2
executing==2.2.1
fastapi==0.126.0
filelock==3.20.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
geojson==3.2.0
giskard==2.18.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.13.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mixpanel==5.0.0
mlflow-skinny==3.7.0
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
num2words==0.5.14
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.14.0
openevals==0.1.3
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
phonenumbers==9.0.21
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
pre_commit==4.5.1
preshed==3.0.12
presidio_analyzer==2.2.360
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyfiglet==1.0.4
Pygments==2.19.2
pyowm==3.5.0
pypdf==6.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
ragas==0.4.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.11.4
sentence-transformers==5.2.0
sentencepiece==0.2.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
sqlparse==0.5.5
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
ty==0.0.6
typer==0.20.1
typer-slim==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
vaderSentiment==3.3.2
virtualenv==20.35.4
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.3
websocket-client==1.9.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
